problems/src/main/java/com/korn/SearchInRotatedSortedArray.java
package com.korn;

/**
 * Solution for 
 * <a href="https://leetcode.com/problems/search-in-rotated-sorted-array/">Search in Rotated Sorted Array</a>
 */
public class SearchInRotatedSortedArray {
    public int search(int[] nums, int target) {
        int left = 0;
        int right = nums.length - 1;

        while (left < right) {
            int middle = left + (right - left) / 2;
            int current = nums[middle];

            if (current == target) {
                return middle;
            }
            
            if (nums[left] <= current) {
                if (current > target && target >= nums[left]) {
                    right = middle - 1;
                } else {
                    left = middle + 1;
                }
            } else {
                if (current < target && target <= nums[right]) {
                    left = middle + 1;
                } else {
                    right = middle - 1;
                }
            }
        }

        return nums[left] == target ? left : -1;
    }
}
